+++ sentinelhub/api/__init__.py
@@ -4,13 +4,15 @@ The part of the package that implements interface with Sentinel Hub services.
 from .batch import (
     BatchCollection,
     BatchRequest,
+    BatchRequestStatus,
+    BatchStatisticalRequest,
     BatchTileStatus,
+    BatchUserAction,
     SentinelHubBatch,
+    SentinelHubBatchStatistical,
     monitor_batch_analysis,
     monitor_batch_job,
 )
-from .batch_base import BatchRequestStatus, BatchUserAction
-from .batch_statistical import BatchStatisticalRequest, SentinelHubBatchStatistical
 from .byoc import ByocCollection, ByocCollectionAdditionalData, ByocCollectionBand, ByocTile, SentinelHubBYOC
 from .catalog import SentinelHubCatalog
 from .fis import FisRequest
