+++ sentinelhub/api/batch/base.py
@@ -5,9 +5,9 @@ from abc import ABCMeta
 from enum import Enum
 from typing import Generic, Iterable, Optional, Sequence, Type, TypeVar, Union
 
-from ..constants import RequestType
-from ..type_utils import Json, JsonDict
-from .base import SentinelHubService
+from ...constants import RequestType
+from ...type_utils import Json, JsonDict
+from ..base import SentinelHubService
 
 BatchRequestType = TypeVar("BatchRequestType", bound="BaseBatchRequest")  # pylint: disable=invalid-name
 RequestSpec = Union[str, dict, BatchRequestType]
