+++ sentinelhub/api/batch/statistical.py
@@ -12,11 +12,11 @@ from dataclasses_json import CatchAll, LetterCase, Undefined
 from dataclasses_json import config as dataclass_config
 from dataclasses_json import dataclass_json
 
-from ..type_utils import Json, JsonDict
-from .base_request import InputDataDict
-from .batch_base import BaseBatchClient, BaseBatchRequest, BatchRequestStatus, BatchUserAction
-from .statistical import SentinelHubStatistical
-from .utils import datetime_config, enum_config, remove_undefined
+from ...type_utils import Json, JsonDict
+from ..base_request import InputDataDict
+from ..statistical import SentinelHubStatistical
+from ..utils import datetime_config, enum_config, remove_undefined
+from .base import BaseBatchClient, BaseBatchRequest, BatchRequestStatus, BatchUserAction
 
 if sys.version_info < (3, 11):
     from typing_extensions import NotRequired, TypedDict
