+++ tests/api/batch/test_statistical.py
